How Sad!
Bugs I can't be bothered to fix right now make me sad. Almost all manual
tests I did before were only with a single segment (as by default), so
this is sad. As said in the README, I'm currently uninterested in fixing
it and will get around to it later in development