Added some initial convolution support
Copied/Rewrote the libcorrect code, no idea if it actually works, for
encoding convolutional streams, and copied libfec into the working tree.
Idea is to statically link libfec into the tree, if not available
according to configure.ac.

After this, I'll add the wrapper functions for these and change the
format, test around to see if it works, and revert this commit if its
uber slow or broken.