Removal of nblock from header
With the removal of segments, we don't need to know how many blocks have
been made, just to reconstruct them faithfully, nbyte by nbyte. Now shit
really won't compile, because I haven't rewritten the
construct/deconstruct functions to support this.

This is ultimately in plan for the format to be fully streamable, with
zero seeks. The ideal set up would be to have the whole pipeline for
sending a compressed tar and untaring it over the web to be something
like this, with no files in between.

[receiver]
nc -l -p 7000 | bef -d | unzstd | tar x

[sender]
tar c dir | zstd | bef | nc receiver 7000