Adds in raw mode; More fixes
Adds in raw mode, which is a low level feature that allows you to skip
the header if you already know the fragment size + other requisite
details. This gets rid of the biggest single source of failure in the
format.

Also fixed pipes (again). Don't know why I'm struggling so much with
them...

Also cleaned up the header a bit.