How Sad! Block Numbers are back!
I, in my naivety, actually removed them (and fragment numbers) a while
back. At least fragment numbers are still gone, but block numbers have
come back because they are necessary for one potentiality that could
arise if there was insertion and/or deletion.

It prevents us from passing a fragment from a different block into the parity
backend, in the event it appears. This is important not just for poor fragments
that may show up at the end of the buffer, but also for the interleaved
fragments since we can no longer assume correct ordering.