Fixed a buffer overflow that failed Alpine tests
I even warned myself about this the function above...

Also added check for whether the scan distance is greater than the
fragment size or not. This is what I get for not testing on Alpine, just
asumming it works like it does on Debian. What a terrible event, I'll
need to release a hotfix v0.2.1...