A couple of changes
Created a few global variables so we don't have to create new contexts
all the time. Also changed padding interface to not zero stuff out.
Stuff from here on out would be to optimize the memory usage, both to
decrease the number of redundant buffers and make it faster by removing
any useless memcpys or memsets, and see if I can actually optimize
zfec.c this time (last attempt as a failure!).