Changed Block Size Test to not gobble memory
We're only testing the block size, not the interleave. Under default
settings, assuming the backend allocates its own memory and we have no
redundancies, that's at least 6G of memory in the final test with 5
blocks interleaved (5G for each block, plus a gig at a time for
processing). Right now it's like twice that, so I can't even test this
block size on 32bit machines.