Actually Don't Buffer Pipes
The reason we did it before was due to our broken read/write interface.
Here it doesn't care if it only returns 64K or whatever the buffer is,
it just increases the offset and keeps on going. Tar and such work just
fine. This makes it easier to port too, as we aren't dependent on how
big the pipe buffer is.