Reapply "Parallelize construction of fragments too"
This reverts commit 39d2b4149a9ebbe944f00045afc91b56b5e9ad7a.

Reverting it again, it appears the problem is that with that many
threads it thrashes the memory bandwidth of the Xeon machine, which with
a fast hash like xxhash, slows stuff down significantly. When having it
set to the prior core count (number of interleaved fragments), it
becomes fast again and has a significant speed up for slower hashes,
though it still slows down xxhash. All in all, I think for a small
performance hit against xxhash here, which recovers at high interleaved
counts, a significant boost for hashes like SHA256 and BLAKE3 is good.