Added Leopard Support
Pretty nice codec, can handle some very large k/m sizes. After this
comes Wirehair and then v0.2.3. Yes, multithreading, a couple of flags,
and a bunch of parity backends is not enough for a bump to v0.3, I need
to add a significant feature for that (such as convolutional coding).