Fix reconstruction again again
We need to keep in mind there may be tail fragments, so if we reach the
end of the loop, but not the end of the bound, we may not discover every
fragment. Note this can't happen with fec-vand and cm256-cauchy, as they
instead rely on replacing erased fragments and thus the bound and the
frag_len should be equal at k.