Change defaults to single threaded
It's uber fast even with a single thread (laptop CPU goes at 1.5GiB/s),
so it's just _wasteful_ to spin all the other threads for that extra
performance, when likely something else is gonna be the bottleneck.
We'll follow with the 0 convention though, and I guess in that route
follow interfaces like zstd that like '-T0' for multithreading.