Buffer Optimization of Deconstruction
Did same thing as with construction. Performance popped up about ~10%
(incredible the difference between the two). From a little bit of
fooling around, it appears the biggest issue is that the checksum
verification is completely serialized (i.e. it can't be parallelized)
due to our scanning algorithm. I wish I knew of a better way to scan for
missin fragments that at least was parallelizable :(. Removing the
fragment verification speeds things like by ~30%, making it a little
faster than construction. So intuitively it must be a significant
bottleneck.

Another signficant bottleneck, also as a result of our scanning
algorithm, is the necessity of having to read ahead on the off chance of
information insertion. This necessitates costly memmoves. On a positive
note, I actually had memcpy there instead of memmove, so this commit
removes that UB.